requirements/common.txt
requests==2.11.1
h5py==2.6.0
Keras==1.1.2
numpy==1.11.3
Pillow==3.4.2
